Making focus branch's history clearer
If switching to a new focus directly from 'nothing', make no commit on
the branch 'nothing'. This help the history of new focus branch clearer.

Fix 'focus to' fails because of 'set -e' when it checks for the
existence of the focus branch. Bash's `if` accepts a command and will
evaluates its exit code without failing.